feat(trading): refactor reward cards, update tooltip
